feat: Add ident edge map to PostCallerAnalysis
This adds to PostCallerAnalysis a map of function idents and how they could potentially call to each other. This is an extension to the prior commit to provide more static analysis data in a minimal format to help with extended analysis (hopefully avoiding the need for a new callgraph generation).