fix: AST switch to binary API
Using the JSON api resulted in project level custom marshallers causing issues in select cases.
Because we want the raw memory state it's best to avoid this logic and instead use our own binary API.

This is also about 20x faster, and with significantly less memory overhead (granted the improvements are only in the cross-process API overhead).