fix: AST fix for conflicts on dot imports, old go versions and empty bodies
This gracefully handles not tracking usage in modules with go versions too old to support our ASTClient. These are reported as an independent value in our coverage.

AST client and API fields and functions have all been made lower case so that they wont conflict with each other in the case of a dot import.

Also handling a niche condition with empty function bodies.